+++ header/baseRandomTest.h
@@ -1,8 +1,8 @@
 //
 // Creator:    http://www.dicelocksecurity.com
-// Version:    vers.5.0.0.1
+// Version:    vers.6.0.0.1
 //
-// Copyright  2008-2011 DiceLock Security, LLC. All rights reserved.
+// Copyright (C) 2008-2012 DiceLock Security, LLC. All rights reserved.
 //
 //                               DISCLAIMER
 //
@@ -111,7 +111,7 @@ namespace DiceLockSecurity {
 			virtual RandomTests GetType(void) {return NotDefined;};
 
 			/// Gets the minimum stream length 
-			virtual unsigned int GetMinimumLength(void) {return 0;};
+			virtual unsigned long int GetMinimumLength(void) {return 0;};
 	};
   }
 }
